requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.10.0=py37hc8dfbb8_1
aiohttp=3.7.3=py37he6e4e01_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py37h395d20d_1001
c-ares=1.17.1=hc929b4f_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
certifi=2020.11.8=py37hecd8cb5_0
cffi=1.14.3=py37h2125817_2
chardet=3.0.4=py37h2987424_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.2.1=py37h3b7a55b_0
freetype=2.10.4=ha233b18_0
google-auth=1.23.0=pyhd8ed1ab_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.31.0=py37h7580e61_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.0.0=pyhd8ed1ab_0
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0.1=hab81aa3_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
lz4-c=1.9.2=h79c402e_3
markdown=3.3.3=pyh9f0ad1d_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
multidict=4.7.5=py37h60d8a13_2
ncurses=6.2=h0a44026_1
ninja=1.10.1=py37h879752b_0
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
oauthlib=3.0.1=py_0
olefile=0.46=py37_0
openssl=1.1.1h=haf1e3a3_0
pandas=1.1.3=py37hb1e8313_0
pillow=8.0.1=py37h5270095_0
pip=20.2.4=py37hecd8cb5_0
protobuf=3.13.0.1=py37hb1e8313_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pysocks=1.7.1=py37h2987424_2
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.0=py3.7_0
pytz=2020.1=py_0
readline=8.0=h1de35cc_0
requests=2.25.0=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
setuptools=50.3.1=py37hecd8cb5_1
six=1.15.0=py37hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
tensorboard=2.4.0=pyhd8ed1ab_0
tensorboard-plugin-wit=1.7.0=pyh9f0ad1d_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.7.0=py37
torchvision=0.8.1=py37_cpu
tqdm=4.51.0=pyhd3eb1b0_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.25.11=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.6.3=py37h4b544eb_0
zipp=3.4.0=py_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
